Create listener.py V0.5
Created a semi-skeleton of the listener / poller used to detect when materials fall below a threshold value of 50g in mass.

Will need to figure out how to pass the instance of the materials table to the function, as well as to confirm the decision to use a poll instead of an event listener.