Modify websocket structure
Modified the websocket structure to make it more generalized, allowing for future scalability.  Created a special return method to send data to a specific websocket, which is what we will use to send low stock data.  By specifying low_stock as the UID of the personal message, we can create a specific channel for the frontend to read low stock messages from.