Update controller-main for Mass Updates
Add functionality to have the Arduino send scale readings via MQTT.

The scale checks whether the value is different from the previous value measured.  If so, the scale pushes the data to the new topic_mass.

On the application side, we will need to create a buffer of size 1 to store the most recent mass reading, so when the backend controller method is signaled, it passes over the current mass value.