Add Shelf ID to MQTT Payloads
Added the shelf ID to the payloads being sent from the MC to the MQTT broker.  The shelf ID will now be hardcoded on the MC side, so the backend can differentiate which shelf the information is being sent from.  For users setting up their Arduino, they will need to modify the SHELF_ID macro near the top of the main class of the MC files.