Update QR Method To Obtain Latest Mass Value
Updated the QR method to obtain the latest mass value from the scale listener.  This class now calls the instance of the MQTTscale object initialized in EmbeddedListener, allowing us to simply extract the value being read by calling the function.  Updated the integration test for this function as well to properly test the new functionality.