Auto Format + Conditional Modification
Auto formatted the main file, and also changed the if to an else if within the main loop for detecting network and mqtt connection.  It may be possible that we try to reconnect to MQTT although already connected, since the WiFi connect function will assume MQTT is disconnected as well and reconnects to it already.