Humidity scaling function
Created a function to model the humidity differences more accurately.  Used a linear function approach, with a high reading of 42% and low reading of 16% RH from the Govee sensor to calibrate this function.

Also added a longer delay to the MQTT reconnect logic in the event of a timeout for many requests to the broker.