Complete Connection of Mass Method to QR Controller
Added the start() function command to the scale_mqtt_instance since it was not starting, this line may have been deleted on accident in a previous commit.

Removing the test_qr_endpoint test, as there's no clean way to decouple features like MQTT for testing in the event the broker is down.